requirements.txt
kelvin-sdk>=8.0
requests>=2.28
keyring>=24.0
